+++ src/cli/commands/tool/prompt/builder.ts
