+++ src/cli/commands/tool/prompt/handler.ts
@@ -1,7 +1,7 @@
 import { ToolCliArgs } from "@/src/types/cli_types.js";
 import { selectTemplate } from "@/src/utils/inquirer.js";
-import { dumpPromptFromTemplate } from "../../../tools/dump_prompt.js";
-import { setGraphAILogger } from "@/src/utils/cli.js";
+import { dumpPromptFromTemplate } from "../../../../tools/dump_prompt.js";
+import { setGraphAILogger } from "@/src/cli/helpers.js";
 
 export const handler = async (argv: ToolCliArgs<{ template?: string }>) => {
   let { template } = argv;
