+++ src/cli/commands/tool/prompt/index.ts
