+++ src/cli/commands/tool/schema/builder.ts
