+++ src/cli/commands/tool/schema/index.ts
