+++ src/cli/commands/tool/scripting/index.ts
